player一次大的结构调整，修改了一大坨代码,已经有点晕了，还没完全调试完，先这个吧

+++ player/playerlib/MessageFFMpegPacket.hpp
@@ -49,13 +49,13 @@ namespace message {
 		//
 		uint32_t mPacketType;
 		//
-		//  是否i桢
+		//  是否i桢,视频有效
 		//
 		int32_t mIFrame;
 		//
 		// 
 		//
-		int32_t mSerialNumber;
+		int64_t mSerialNumber;
 	};
 
 }
